#125 Refactor Game and Event view model methods
